Configuração do banco firebase
