+++ src/leio/Components/Home.js
@@ -4,9 +4,6 @@ import { View, Text } from "react-native";
 import styles from "../assets/styles/base";
 
 const Home = () => {
-  //Para testar hover no botão
-  //const [isHovered, setIsHovered] = useState(false);
-
   return (
     <View style={styles.container}>
       <Text>Apenas um teste</Text>
