--- src/leio/firebaseConfig.js
@@ -1,15 +1,0 @@
-// Import the functions you need from the SDKs you need
-import { initializeApp } from "firebase/app";
-import { getFirestore, collection, getDocs } from "firebase/firestore/lite";
-// TODO: Add SDKs for Firebase products that you want to use
-// https://firebase.google.com/docs/web/setup#available-libraries
-
-// Your web app's Firebase configuration
-const firebaseConfig = {
-  apiKey: "AIzaSyC7bY3LOXNFzvM86jPSQyG5i6L3Nnf4X_0",
-  authDomain: "leio-app.firebaseapp.com",
-  projectId: "leio-app",
-  storageBucket: "leio-app.appspot.com",
-  messagingSenderId: "1077605673545",
-  appId: "1:1077605673545:web:7552bb30f32699d325ed98",
-};
